rename to `workorder-details`, add a changelog entry
