+++ changelog.txt
@@ -15,6 +15,8 @@ that repo.
 
 ## New Scripts
 
+- `gui/workorder-details`: adjusts work orders' input item, material, traits
+
 ## Removed
 
 ## Fixes
